refactor: consolidate duplicated test fixtures across admin test files
The five admin test modules (test_admin, test_admin_auth, test_admin_config,
test_admin_moderation, test_admin_reports) each carried their own copies of
the settings-cache shim, mock_neo4j, and the admin/noauth/regular app+client
fixtures — with the latter four importing helpers from test_admin, coupling
the files together. Moved the shared fixtures and the _admin_user /
_regular_user / _test_settings helpers into tests/test_api/conftest.py so the
admin suite has a single source.

_test_settings() keeps AuthSettings() defaults (no user_service_url /
user_service_jwks_cache_ttl) — that distinction from tests/test_auth/conftest.py
is intentional and preserved; the two settings fixtures are not merged.

test_admin_config.py keeps a local mock_pg + get_pg override since that is
specific to the config endpoints.

No tests added or removed: 221 test_api+test_auth tests pass (same as before).

Closes #795
