+++ tests/test_api/test_admin.py
@@ -1,4 +1,11 @@
-"""Tests for admin API endpoints."""
+"""Tests for admin API endpoints.
+
+Shared fixtures (``mock_neo4j``, ``admin_app``/``admin_client``,
+``noauth_app``/``noauth_client``, ``regular_app``/``regular_client``, the
+``_clear_settings_cache`` autouse shim) and the ``_admin_user`` /
+``_regular_user`` / ``_test_settings`` helpers live in
+``tests/test_api/conftest.py``.
+"""
 
 from __future__ import annotations
 
@@ -7,119 +14,6 @@ from unittest.mock import MagicMock, patch
 import pytest
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
-
-from src.api.auth import User
-from src.api.middleware import require_admin
-from src.config import (
-    AuthSettings,
-    Neo4jSettings,
-    PostgresSettings,
-    RedisSettings,
-    Settings,
-    get_settings,
-)
-
-
-def _admin_user() -> User:
-    return User(
-        id="admin-user",
-        email="admin@example.com",
-        name="Admin User",
-        is_admin=True,
-    )
-
-
-def _regular_user() -> User:
-    return User(
-        id="regular-user",
-        email="user@example.com",
-        name="Regular User",
-        is_admin=False,
-    )
-
-
-def _test_settings() -> Settings:
-    """Build a Settings instance without reading .env."""
-    return Settings(
-        _env_file=None,
-        neo4j=Neo4jSettings(uri="bolt://localhost:7687", user="neo4j", password="test"),
-        postgres=PostgresSettings(dsn="postgresql://test:test@localhost:5432/test"),
-        redis=RedisSettings(url="redis://localhost:6379/0"),
-        auth=AuthSettings(),
-    )
-
-
-@pytest.fixture(autouse=True)
-def _clear_settings_cache(monkeypatch: pytest.MonkeyPatch) -> None:
-    """Patch get_settings to avoid .env parsing errors in tests."""
-    test_settings = _test_settings()
-    get_settings.cache_clear()
-
-    import src.config
-
-    monkeypatch.setattr(src.config, "get_settings", lambda: test_settings)
-
-
-@pytest.fixture
-def mock_neo4j() -> MagicMock:
-    client = MagicMock()
-    client.execute_read.return_value = []
-    client.execute_write.return_value = []
-    return client
-
-
-@pytest.fixture
-def admin_app(mock_neo4j: MagicMock) -> FastAPI:
-    """FastAPI app with admin auth override."""
-    from src.api.app import create_app
-
-    app = create_app()
-    app.state.neo4j = mock_neo4j
-    app.dependency_overrides[require_admin] = _admin_user
-    return app
-
-
-@pytest.fixture
-def admin_client(admin_app: FastAPI) -> TestClient:
-    return TestClient(admin_app)
-
-
-@pytest.fixture
-def noauth_app(mock_neo4j: MagicMock) -> FastAPI:
-    """FastAPI app with NO auth overrides — tests 401/403 paths."""
-    from src.api.app import create_app
-
-    app = create_app()
-    app.state.neo4j = mock_neo4j
-    return app
-
-
-@pytest.fixture
-def noauth_client(noauth_app: FastAPI) -> TestClient:
-    return TestClient(noauth_app)
-
-
-@pytest.fixture
-def regular_app(mock_neo4j: MagicMock) -> FastAPI:
-    """FastAPI app with non-admin user override for require_admin."""
-    from src.api.app import create_app
-
-    app = create_app()
-    app.state.neo4j = mock_neo4j
-
-    from fastapi import HTTPException
-
-    def _raise_forbidden() -> User:
-        raise HTTPException(status_code=403, detail="Admin access required")
-
-    app.dependency_overrides[require_admin] = _raise_forbidden
-    return app
-
-
-@pytest.fixture
-def regular_client(regular_app: FastAPI) -> TestClient:
-    return TestClient(regular_app)
-
 
 # --- Health endpoints ---
